+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Enable client-side round-robin DNS load balancing for all gRPC connections.
 
 ## [1.0.0-alpha.1] - 2021-09-28
 ### Added